Added background check for host information
New modules that checks the host information database for
completeness. This module requires an oracle database which
contains at least the field "IP".
The module will search the srcIP field in the flow-database
for IP address and checks all seen IPs against the host
information database. Alarms are raised whenever an IP
was active (e.g. emitted flows) and is not documented in the
host information database.