Do not fetch data before interval is defined
Several views, e.g. Dashboard, want want a time interval in order
to fetch and show its data. The system fetched information about
the complete data set when no interval was defined, because the
TimelineModel did not have information about the available bucket
sizes. The complete data is immediately dropped after the minimum
time interval is known by the timeline. Thus, fetching the complete
data set puts unnecessary load onto the database and increases
waiting time before the page is completely rendered.

This commit introduces the "fetchEmptyInterval" parameter to the views
which can be set by the page. If the value of this variable is set  to
false, the views will not fetch any data until an interval has been
defined for their models. The default value for "fetchEmptyInterval"
is true, thus preserving the old behaviour for all views which do  not
explicitly set the new parameter.