Bugfix: SQL backends and total counters
The total counter in the sql backends was encoded as IP address
0. This works unless you have DHCP traffic in your flow data
which regularly uses the IP address 0 as a source IP address.
This behavior first screws with the total counter, and could lead
to bugs in the EdgeBundle as there were flows that contained the
IP address 0 but there was no node with this id (since IP 0 was
reserved for the total counters).

This commit fixes the problem by making a change to the database
layout and encoding the total counter into a special id. It is
required to rebuild the flow database in order for this version
of flow-inspector to be able to work. You cannot reuse your old
databases.